+++ src/commands/leaderboard.ts
@@ -8,6 +8,8 @@ import {
 } from 'discord.js';
 import CitizenModel from '../db/models/citizen';
 
+type LeaderboardType = 'dings' | 'gulags';
+
 @Discord()
 @SlashGroup('leaderboard')
 export default class Leaderboard {
@@ -17,41 +19,54 @@ export default class Leaderboard {
     description: 'See who has talked with their mic muted the most (cross-server)',
   })
   async leaderboardDings(interaction: CommandInteraction, client: Client): Promise<void> {
-    const options = await this.generateLeaderboardOptions(client);
+    const options = await this.generateLeaderboardOptions(client, 'dings');
     await interaction.reply(options);
   }
 
-  @ButtonComponent('prev-dings-btn')
-  async prevDingsBtn(interaction: ButtonInteraction, client: Client): Promise<void> {
+  @Slash('gulags', {
+    description: 'See who has been sent to the gulag the most (cross-server)',
+  })
+  async leaderboardGulag(interaction: CommandInteraction, client: Client): Promise<void> {
+    const options = await this.generateLeaderboardOptions(client, 'gulags');
+    await interaction.reply(options);
+  }
+
+  @ButtonComponent('prev-leaderboard-btn')
+  async prevLeaderboardBtn(interaction: ButtonInteraction, client: Client): Promise<void> {
+    const { title } = interaction.message.embeds[0];
+    const type = title.includes('dings') ? 'dings' : 'gulags';
     const [, from] = interaction.message.embeds[0].description.match(/Top (\d+)-(\d+)/);
     const skip = Number(from) - (this.limit + 1);
-    const options = await this.generateLeaderboardOptions(client, skip >= 0 ? skip : 0);
+    const options = await this.generateLeaderboardOptions(client, type, skip >= 0 ? skip : 0);
     await interaction.update(options);
   }
 
-  @ButtonComponent('next-dings-btn')
-  async nextDingsBtn(interaction: ButtonInteraction, client: Client): Promise<void> {
+  @ButtonComponent('next-leaderboard-btn')
+  async nextLeaderboardBtn(interaction: ButtonInteraction, client: Client): Promise<void> {
+    const { title } = interaction.message.embeds[0];
+    const type = title.includes('dings') ? 'dings' : 'gulags';
     const [, , to] = interaction.message.embeds[0].description.match(/Top (\d+)-(\d+)/);
     const skip = Number(to);
-    const options = await this.generateLeaderboardOptions(client, skip);
+    const options = await this.generateLeaderboardOptions(client, type, skip);
     await interaction.update(options);
   }
 
-  async generateLeaderboardOptions(client: Client, skip = 0) {
-    const topDings = await CitizenModel.find({ dings: { $gt: 0 } })
-      .sort({ dings: -1 })
+  async generateLeaderboardOptions(client: Client, type: LeaderboardType, skip = 0) {
+    const key = type === 'dings' ? 'dings' : 'gulagCount';
+    const page = await CitizenModel.find({ [key]: { $gt: 0 } })
+      .sort({ [key]: -1 })
       .limit(this.limit)
       .skip(skip);
 
     const embed = new MessageEmbed({
-      title: 'Top dings globally',
-      description: `Top ${skip + 1}-${skip + topDings.length}`,
+      title: `Top ${type} globally`,
+      description: `Top ${skip + 1}-${skip + page.length}`,
       fields: await Promise.all(
-        topDings.map(async (citizen, i) => ({
+        page.map(async (citizen, i) => ({
           name: `${i + skip + 1}: ${
             (client.users.cache.get(citizen._id) || (await client.users.fetch(citizen._id))).tag
           }`,
-          value: citizen.dings.toString(),
+          value: citizen[key].toString(),
           inline: false,
         }))
       ),
@@ -61,13 +76,13 @@ export default class Leaderboard {
       .setEmoji('⏮')
       .setStyle('PRIMARY')
       .setDisabled(!skip)
-      .setCustomId('prev-dings-btn');
+      .setCustomId('prev-leaderboard-btn');
     const nextBtn = new MessageButton()
       .setLabel('Next')
       .setEmoji('⏭')
       .setStyle('PRIMARY')
-      .setDisabled((await CitizenModel.count({ dings: { $gt: 0 } })) <= skip + this.limit)
-      .setCustomId('next-dings-btn');
+      .setDisabled((await CitizenModel.count({ [key]: { $gt: 0 } })) <= skip + this.limit)
+      .setCustomId('next-leaderboard-btn');
 
     return {
       embeds: [embed],